chore: tidy up onRunJourney
