+++ electron/execution.ts
@@ -22,246 +22,22 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 */
 
-import { join } from 'path';
-import { writeFile, rm, mkdir } from 'fs/promises';
 import { EventEmitter } from 'events';
-import { shell, BrowserWindow, ipcMain, IpcMainInvokeEvent } from 'electron';
-import { fork, ChildProcess } from 'child_process';
-import logger from 'electron-log';
-import isDev from 'electron-is-dev';
-import { JOURNEY_DIR, PLAYWRIGHT_BROWSERS_PATH } from './config';
-import type {
-  ActionInContext,
-  RecorderSteps,
-  RunJourneyOptions,
-  StepEndEvent,
-  StepStatus,
-  TestEvent,
-} from '../common/types';
-import { SyntheticsGenerator } from './syntheticsGenerator';
-import { BrowserManager, browserManager } from './browserManager';
-import { onRecordJourneys, onSetMode, onExportScript } from './api';
+import { ipcMain, IpcMainInvokeEvent } from 'electron';
+import type { RunJourneyOptions } from '../common/types';
+import { browserManager } from './browserManager';
+import {
+  onSetMode,
+  onExportScript,
+  runJourney,
+  recordJourney,
+  onOpenExternalLink,
+  onGenerateCode,
+} from './api';
+import { syntheticsManager } from './syntheticsManager';
 
-const SYNTHETICS_CLI = require.resolve('@elastic/synthetics/dist/cli');
-
-// TODO: setting isBrowserRunning from onRecordJourney is broken
 export enum MainWindowEvent {
   MAIN_CLOSE = 'main-close',
-}
-
-/**
- * Attempts to find the step associated with a `step/end` event.
- *
- * If the step is found, the sequential titles of each action are overlayed
- * onto the object.
- * @param {*} steps list of steps to search
- * @param {*} event the result data from Playwright
- * @returns the event data combined with action titles in a new object
- */
-function addActionsToStepResult(steps: RecorderSteps, event: StepEndEvent): TestEvent {
-  const step = steps.find(
-    s =>
-      s.actions.length &&
-      s.actions[0].title &&
-      (event.data.name === s.actions[0].title || event.data.name === s.name)
-  );
-  if (!step) return { ...event, data: { ...event.data, actionTitles: [] } };
-  return {
-    ...event,
-    data: {
-      ...event.data,
-      actionTitles: step.actions.map(
-        (action: ActionInContext, index: number) => action?.title ?? `Action ${index + 1}`
-      ),
-    },
-  };
-}
-
-function onTest(browserManager: BrowserManager) {
-  return async function (_event: IpcMainInvokeEvent, data: RunJourneyOptions) {
-    if (browserManager.isRunning()) {
-      throw new Error(
-        'Cannot start testing a journey, a browser operation is already in progress.'
-      );
-    }
-    // TODO: connect onTest with browserManager and refactor
-    // browserManager.isRunning() = true;
-    const parseOrSkip = (chunk: string): Array<Record<string, any>> => {
-      // at times stdout ships multiple steps in one chunk, broken by newline,
-      // so here we split on the newline
-      return chunk.split('\n').map(subChunk => {
-        try {
-          return JSON.parse(subChunk);
-        } catch (_) {
-          return {};
-        }
-      });
-    };
-    const isJourneyStart = (event: any): event is { journey: { name: string } } => {
-      return event.type === 'journey/start' && !!event.journey.name;
-    };
-
-    const isStepEnd = (
-      event: any
-    ): event is {
-      step: { duration: { us: number }; name: string; status: StepStatus };
-      error?: Error;
-    } => {
-      return (
-        event.type === 'step/end' &&
-        ['succeeded', 'failed', 'skipped'].includes(event.step?.status) &&
-        typeof event.step?.duration?.us === 'number'
-      );
-    };
-
-    const isJourneyEnd = (
-      event: any
-    ): event is { journey: { name: string; status: 'succeeded' | 'failed' } } => {
-      return (
-        event.type === 'journey/end' && ['succeeded', 'failed'].includes(event.journey?.status)
-      );
-    };
-
-    const constructEvent = (parsed: Record<string, any>): TestEvent | null => {
-      if (isJourneyStart(parsed)) {
-        return {
-          event: 'journey/start',
-          data: {
-            name: parsed.journey.name,
-          },
-        };
-      }
-      if (isStepEnd(parsed)) {
-        return {
-          event: 'step/end',
-          data: {
-            name: parsed.step.name,
-            status: parsed.step.status,
-            duration: Math.ceil(parsed.step.duration.us / 1000),
-            error: parsed.error,
-          },
-        };
-      }
-      if (isJourneyEnd(parsed)) {
-        return {
-          event: 'journey/end',
-          data: {
-            name: parsed.journey.name,
-            status: parsed.journey.status,
-          },
-        };
-      }
-      return null;
-    };
-    // TODO: de-deup browserWindow getter
-    const browserWindow = BrowserWindow.getFocusedWindow() || BrowserWindow.getAllWindows()[0];
-    const sendTestEvent = (event: TestEvent) => {
-      browserWindow.webContents.send('test-event', event);
-    };
-
-    const emitResult = (chunk: string) => {
-      parseOrSkip(chunk).forEach(parsed => {
-        const event = constructEvent(parsed);
-        if (event) {
-          sendTestEvent(
-            event.event === 'step/end' ? addActionsToStepResult(data.steps, event) : event
-          );
-        }
-      });
-    };
-
-    let synthCliProcess: ChildProcess | null = null; // child process, define here to kill when finished
-
-    try {
-      const isProject = data.isProject;
-      const args = [
-        '--no-headless',
-        '--reporter=json',
-        '--screenshots=off',
-        '--no-throttling',
-        '--sandbox',
-      ];
-      const filePath = join(JOURNEY_DIR, 'recorded.journey.js');
-      if (!isProject) {
-        args.push('--inline');
-      } else {
-        await mkdir(JOURNEY_DIR, { recursive: true });
-        await writeFile(filePath, data.code);
-        args.unshift(filePath);
-      }
-
-      /**
-       * Fork the Synthetics CLI with correct browser path and
-       * cwd correctly spawns the process
-       */
-      synthCliProcess = fork(`${SYNTHETICS_CLI}`, args, {
-        env: {
-          ...process.env,
-          PLAYWRIGHT_BROWSERS_PATH,
-        },
-        cwd: isDev ? process.cwd() : process.resourcesPath,
-        stdio: 'pipe',
-      });
-
-      function handleMainClose() {
-        if (synthCliProcess && !synthCliProcess.kill()) {
-          logger.warn('Unable to abort Synthetics test proceess.');
-        }
-      }
-      // mainWindowEmitter.addListener(MainWindowEvent.MAIN_CLOSE, handleMainClose);
-
-      const { stdout, stdin, stderr } = synthCliProcess as ChildProcess;
-      if (!isProject) {
-        stdin?.write(data.code);
-        stdin?.end();
-      }
-      stdout?.setEncoding('utf-8');
-      stderr?.setEncoding('utf-8');
-      for await (const chunk of stdout!) {
-        emitResult(chunk);
-      }
-      for await (const chunk of stderr!) {
-        logger.error(chunk);
-      }
-      if (isProject) {
-        await rm(filePath, { recursive: true, force: true });
-      }
-
-      // mainWindowEmitter.removeListener(MainWindowEvent.MAIN_CLOSE, handleMainClose);
-    } catch (error: unknown) {
-      logger.error(error);
-      sendTestEvent({
-        event: 'journey/end',
-        data: {
-          status: 'failed',
-          error: error as Error,
-        },
-      });
-    } finally {
-      if (synthCliProcess && !synthCliProcess.kill()) {
-        logger.warn(
-          `Attempted to send SIGTERM to synthetics process, but did not receive exit signal. Process ID is ${synthCliProcess.pid}.`
-        );
-      }
-      // isBrowserRunning = false;
-    }
-  };
-}
-
-async function onGenerateCode(
-  _event: IpcMainInvokeEvent,
-  data: { isProject: boolean; actions: RecorderSteps }
-) {
-  const generator = new SyntheticsGenerator(data.isProject);
-  return generator.generateFromSteps(data.actions);
-}
-
-async function onLinkExternal(_event: IpcMainInvokeEvent, url: string) {
-  try {
-    await shell.openExternal(url);
-  } catch (e) {
-    logger.error(e);
-  }
 }
 
 /**
@@ -278,13 +54,34 @@ export default function setupListeners(mainWindowEmitter: EventEmitter) {
     if (browserManager.isRunning()) {
       await browserManager.closeBrowser();
     }
+
+    if (syntheticsManager.isRunning()) {
+      await syntheticsManager.stop();
+    }
   });
-  ipcMain.handle('record-journey', onRecordJourneys(browserManager));
-  ipcMain.handle('run-journey', onTest(browserManager));
+
+  ipcMain.handle('record-journey', onRecordJourney);
+  ipcMain.handle('run-journey', onRunJourney);
   ipcMain.handle('actions-to-code', onGenerateCode);
   ipcMain.handle('export-script', onExportScript);
   ipcMain.handle('set-mode', onSetMode(browserManager));
-  ipcMain.handle('link-to-external', onLinkExternal);
+  ipcMain.handle('open-external-link', onOpenExternalLink);
 
   return () => ipcMain.removeAllListeners();
+}
+
+async function onRecordJourney(event: IpcMainInvokeEvent, url: string) {
+  if (browserManager.isRunning() || syntheticsManager.isRunning()) {
+    throw new Error(
+      'Cannot start recording a journey, a browser operation is already in progress.'
+    );
+  }
+  await recordJourney(event, url, browserManager);
+}
+
+async function onRunJourney(event: IpcMainInvokeEvent, data: RunJourneyOptions) {
+  if (browserManager.isRunning() || syntheticsManager.isRunning()) {
+    throw new Error('Cannot start testing a journey, a browser operation is already in progress.');
+  }
+  await runJourney(event, data, syntheticsManager);
 }
