+++ electron/syntheticsManager.ts
@@ -21,21 +21,33 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 */
-import { IpcMainInvokeEvent } from 'electron';
-import { BrowserManager } from '../browserManager';
+import { fork, ChildProcess, ForkOptions } from 'child_process';
+import logger from 'electron-log';
+const SYNTHETICS_CLI = require.resolve('@elastic/synthetics/dist/cli');
 
-export function onSetMode(browserManager: BrowserManager) {
-  return async function (_event: IpcMainInvokeEvent, mode: string) {
-    const browserContext = browserManager.getContext();
-    if (!browserContext) return;
-    const page = browserContext.pages()[0];
-    if (!page) return;
-    await page.mainFrame().evaluate(
-      ([mode]) => {
-        // `__pw_setMode` is a private function
-        (window as any).__pw_setMode(mode);
-      },
-      [mode]
-    );
-  };
+export class SyntheticsManager {
+  protected _cliProcess: ChildProcess | null = null;
+
+  isRunning() {
+    return !!this._cliProcess;
+  }
+
+  /**
+   * Fork the Synthetics CLI with correct browser path and
+   * cwd correctly spawns the process
+   */
+  run(args: string[], options: ForkOptions) {
+    const ps = fork(`${SYNTHETICS_CLI}`, args, options);
+    this._cliProcess = ps;
+    return ps;
+  }
+
+  stop() {
+    if (this._cliProcess && !this._cliProcess.kill()) {
+      logger.warn('Unable to abort Synthetics test process.');
+    }
+    this._cliProcess = null;
+  }
 }
+
+export const syntheticsManager = new SyntheticsManager();
